+++ changelog.md
@@ -3,6 +3,7 @@
 ## Next
 
 * Fix: `Promise.coroutine.addYieldHandler()` is maintained
+* Patch `.tapCatch()` prototype method
 * Travis CI run tests on Node versions 8 + 9
 * Skip Travis CI runs on release tags
 * Test against latest Bluebird v3